backend/requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-Mail==0.9.1
python-socketio==5.9.0
python-engineio==4.7.1
bcrypt==4.0.1
PyJWT==2.8.0
python-dotenv==1.0.0
google-generativeai==0.3.2
Werkzeug==2.3.7
eventlet==0.33.3
dnspython==2.4.2
email-validator==2.1.0